Updated code, eeprom saving, and sample and slide.
Updated code for the SITH, with eeprom saving of which function was last used. Aswell as a new alternate function Sample and Slide. Also did the changing of alternate codes easier to add more alternate codes and understand it. Also started working in TABS to work with each function more easily.